Cleanup Mastodon+Entity+StatusEdit (IOS-175)
